Task: Delete commented code and console.logs
