+++ src/app/app.component.html
@@ -1,4 +1,3 @@
 <app-banner></app-banner>
-<!--<app-layout></app-layout>-->
 
 <router-outlet></router-outlet>
