+++ src/app/app.component.ts
@@ -1,8 +1,8 @@
 import {ChangeDetectionStrategy, Component} from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 import {HttpClientModule, provideHttpClient} from '@angular/common/http';
-import { BannerComponent } from './banner/banner.component';
-import { ContentModule } from './content/content.module';
+import { BannerComponent } from './components/banner/banner.component';
+import { ContentModule } from './modules/content/content.module';
 
 @Component({
   selector: 'app-root',
@@ -14,7 +14,7 @@ import { ContentModule } from './content/content.module';
   // changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class AppComponent {
-  title = 'crew-app';
+  title = 'start-app';
   constructor() {
     console.log('Crew App');
   }
