+++ src/app/constants/paths.ts
@@ -1,0 +1,10 @@
+const STATS = 'stats';
+
+export const ID = 'id';
+export const CREW = 'crew';
+export const PROFILE = 'profile';
+export const STARLINK = 'starlink';
+export const STARLINK_STATS = `${STARLINK}-${STATS}`;
+export const LAUNCHES = 'launches';
+export const PAYLOADS = 'payloads';
+export const CORES = 'cores';
