--- src/app/content/components/crew/crew.component.scss
@@ -1,10 +1,0 @@
-@import '../../../../styles/colors';
-
-.button-container {
-  text-align: center;
-
-  button:active {
-    color: $button-primary-fg;
-    background-color: $button-primary-bg;
-  }
-}
