--- src/app/content/components/starlink-details/starlink-details.component.scss
@@ -1,17 +1,0 @@
-@import '../../../../styles/fonts';
-@import '../../../../styles/colors';
-
-.container {
-  padding: $spacing-2xl $spacing-6xl;
-  border: 1px solid $border-secondary;
-  border-radius: $radius-xl;
-  margin: $spacing-2xl;
-  //border-radius: $radius-md;
-  //border: 1px solid $border-secondary;
-  //background-color: $alpha-white-100;
-  //box-shadow: $shadow-lg;
-}
-
-h3 {
-  text-align: center;
-}
