--- src/app/content/components/starlink-stats/starlink-stats.component.html
@@ -1,38 +1,0 @@
-<div *ngIf="loading" class="loader">Loading...</div>
-
-<div *ngIf="!loading" class="container">
-  <h3>Star Links</h3>
-
-  {{ lists.length }}
-  <div *ngIf="!areAllListsEmpty; else noDate" class="lists-container">
-    <div *ngFor="let list of lists; let index = index" class="list">
-      <div>
-        <h3>List {{ index + 1 }}</h3>
-
-        <div *ngFor="let starlink of list" class="list-item" [routerLink]="['/starlink-stats', starlink.launch]">
-          <p>id: {{ starlink.id }}</p>
-
-          <p>creation date: {{ starlink.creationDate | dateFormat }}</p>
-
-          <p>country code: {{ starlink.countryCode }}</p>
-
-          <p>height (km): {{ starlink.height_km }}</p>
-        </div>
-      </div>
-    </div>
-  </div>
-
-  <ng-template #noDate class="no-data">
-    <p>No data available</p>
-  </ng-template>
-</div>
-
-<div class="navigation">
-  <div *ngIf="crewMemberId">
-    <a [routerLink]="['/profile', crewMemberId]">Profile</a>
-  </div>
-
-  <div>
-    <a [routerLink]="'/'">Leave Profile</a>
-  </div>
-</div>
