--- src/app/content/constants/urls.ts
@@ -1,6 +1,0 @@
-const API_URL = 'https://api.spacexdata.com/v4';
-export const CREW_MEMBER_URl = `${API_URL}/crew`;
-export const STAR_LINK_URL = `${API_URL}/starlink`;
-export const LAUNCH_URL = `${API_URL}/launches`;
-export const PAYLOAD_URL = `${API_URL}/payloads`;
-export const CORE_URL = `${API_URL}/cores`;
