--- src/app/layout/layout.component.html
@@ -1,3 +1,0 @@
-<app-banner></app-banner>
-
-<router-outlet></router-outlet>
