+++ src/app/mocks/crew-member-ids.ts
@@ -1,0 +1,5 @@
+export const predefinedCrewMemberIds = ['5ebf1a6e23a9a60006e03a7a',
+  '5ebf1b7323a9a60006e03a7b',
+  '5f7f1543bf32c864a529b23e',
+  '5f7f158bbf32c864a529b23f',
+  '5f7f15d5bf32c864a529b240'];
