+++ src/app/mocks/upcoming-launches-data.ts
@@ -1,0 +1,5 @@
+export const upcomingLaunchesData = {
+  name: 'Apollo 24',
+  step: 'Health check - pending',
+  progress: 60,
+};
