+++ src/app/modules/content/components/starlink-stats/starlink-stats.component.html
@@ -1,0 +1,39 @@
+<div *ngIf="loading" class="loader">Loading...</div>
+
+<div *ngIf="!loading" class="container">
+  <h3>Star Links</h3>
+
+  <div *ngIf="!areAllListsEmpty; else noData" class="lists-container">
+    <div *ngFor="let list of lists; let index = index" class="list">
+      <div>
+        <h3>List {{ index + 1 }}</h3>
+
+        <div *ngFor="let starlink of list" class="list-item" [routerLink]="['/starlink-stats', starlink.launch]">
+          <p>id: {{ starlink.id }}</p>
+
+          <p>creation date: {{ starlink.creationDate | dateFormat }}</p>
+
+          <p>country code: {{ starlink.countryCode }}</p>
+
+          <p>height (km): {{ starlink.height_km }}</p>
+        </div>
+      </div>
+    </div>
+  </div>
+
+  <ng-template #noData>
+    <div class="no-data">
+      <p>No data available</p>
+    </div>
+  </ng-template>
+</div>
+
+<div class="navigation">
+  <div *ngIf="crewMemberId">
+    <a [routerLink]="['/profile', crewMemberId]">Profile</a>
+  </div>
+
+  <div>
+    <a [routerLink]="'/'">Leave Profile</a>
+  </div>
+</div>
